+++ tests/testthat/test-get_name.R
@@ -7,3 +7,7 @@ test_that("works for eq objects", {
 test_that("works for eqs objects", {
   expect_equal(get_name(eqs_test), c("eqs-test"))
 })
+
+test_that("works for eqs_bag objects", {
+  expect_equal(get_name(eqs_bag_test), c("overall-bag"))
+})