+++ tests/testthat/test-get_outcome.R
@@ -5,6 +5,10 @@ test_that("works for eq objects", {
 })
 
 
-test_that("works for eq objects", {
+test_that("works for eqs objects", {
   expect_equal(get_outcome(eqs_test), rep("kcal/day", 4))
+})
+
+test_that("works for eqs_bag objects", {
+  expect_equal(get_outcome(eqs_bag_test), c("kcal/day", "kcal/month"))
 })